Fix https://github.com/NuGet/Home/issues/225 Fix https://github.com/NuGet/Home/issues/1547
1. Change the order of uninstall to run the uninstall.ps1 script first (to match NuGet 2)
2. Check that the folder exists before trying to delete it
3. Remove duplicate call to Remove file from the project
